+++ client_test.go
@@ -1,7 +1,6 @@
 package honeybadger
 
 import (
-	"sync"
 	"testing"
 )
 
@@ -27,48 +26,6 @@ func TestConfigureClientEndpoint(t *testing.T) {
 	if *backend.URL != "http://localhost:3000" {
 		t.Errorf("Expected Configure to update backend. expected=%#v actual=%#v", "http://localhost:3000", backend.URL)
 	}
-}
-
-func TestClientContext(t *testing.T) {
-	client := New(Configuration{})
-
-	client.SetContext(Context{"foo": "bar"})
-	client.SetContext(Context{"bar": "baz"})
-
-	context := client.context.internal
-
-	if context["foo"] != "bar" {
-		t.Errorf("Expected client to merge global context. expected=%#v actual=%#v", "bar", context["foo"])
-	}
-
-	if context["bar"] != "baz" {
-		t.Errorf("Expected client to merge global context. expected=%#v actual=%#v", "baz", context["bar"])
-	}
-}
-
-func TestClientConcurrentContext(t *testing.T) {
-	var wg sync.WaitGroup
-
-	client := New(Configuration{})
-	newContext := Context{"foo": "bar"}
-
-	wg.Add(2)
-
-	go updateContext(&wg, client, newContext)
-	go updateContext(&wg, client, newContext)
-
-	wg.Wait()
-
-	context := client.context.internal
-
-	if context["foo"] != "bar" {
-		t.Errorf("Expected context value. expected=%#v result=%#v", "bar", context["foo"])
-	}
-}
-
-func updateContext(wg *sync.WaitGroup, client *Client, context Context) {
-	client.SetContext(context)
-	wg.Done()
 }
 
 func TestNotifyPushesTheEnvelope(t *testing.T) {
@@ -121,9 +78,8 @@ func mockClient(c Configuration) (Client, *mockWorker, *mockBackend) {
 	backendConfig.update(&c)
 
 	client := Client{
-		Config:  newConfig(*backendConfig),
-		worker:  worker,
-		context: newContextSync(),
+		Config: newConfig(*backendConfig),
+		worker: worker,
 	}
 
 	return client, worker, backend